CONCAT specifies optional output colum names
Often, the relations that form the input to a CONCAT have the same
column names, but this is a convention, not a hard rule. Previously, we
were simply using the column names from the first input for the output.
This is still the default behaviour, but it is now possible to override
the column names with a final, optional argument when creating a CONCAT
node.