Pass config object to code generators
This adds a standard CodeGenConfig object that all code generators
expect as an argument to their constructor. We can use this to configure
the input/output directories (hitherto hard-coded) and the delimiter to
use (hitherto hard-coded to "\t" for inputs and "," for outputs), as
well as other configuration options needed (e.g., the Sharemind PID).

This change touches a lot of code as it changes a widely-used interface;
I believe I've fixed all breakage, however (including already broke
examples). @n1v0lg, @bengetch -- please take a look.