Use separate generated code directory
Previously, we were storing the generated code either in the input or in
the output directory in an ad-hoc way. This is problematic when the
input/output directories are on HDFS. Hence, this change introduces a
separate `code_path` configuration entry that sets the path for
generated code. If not specified, we choose a random, unique temporary
directory.