Collect Sharemind input data from HDFS
This concatenates any HDFS input files into a single local file
for Sharemind to use, and prepends the data with a header describing
the columns (required by Sharemind).

Note that this has some caveats currently: it won't work with when
running with local-only Spark, and it doesn't support Sharemind jobs
running directly on input data (as that won't have the right layout
in HDFS).